+++ NEWS.md
@@ -2,9 +2,13 @@ General
 
 - PCA improvements + associated vignettes
 - scRNA integration assessment metrics
+- GSEA barplot + associated helper functions
 
 New
 
+- `plot_GSEA_barplot`: function-ified version of horizontal barplot originally made for Jack
+  - `format_GSEA_name`: cleans up MSigDB underscore + all caps names a bit
+  - `split_line`: splits long text to multiple lines based on number of char or lines
 - `rotate_varimax`: Varimax rotation for prcomp output
 - `assess_integration`: `run_LISI` but more general, incorporating `Seurat::MixingMetric`
   - `run_LISI`: wrapper for `assess_integration` specific to `LISI`